+++ universim2/moon.cpp
@@ -1,11 +1,11 @@
 #include "moon.hpp"
 
-Moon::Moon(const char *name, double radius, double mass, double meanDistance, double eccentricity, double inclination) : StellarObject(name, 4, radius, mass, meanDistance, eccentricity, inclination){
+Moon::Moon(const char *name, long double radius, long double mass, long double meanDistance, long double eccentricity, long double inclination) : StellarObject(name, 4, radius, mass, meanDistance, eccentricity, inclination){
     // ------------------------------------------------- TODO -------------------------------------------------
     // Set random colour
     setColour(0x964B00);
 }
 
-Moon::Moon(const char *name, double radius, double mass, double meanDistance, double eccentricity, double inclination, int colour) : StellarObject(name, 4, radius, mass, meanDistance, eccentricity, inclination, colour){
+Moon::Moon(const char *name, long double radius, long double mass, long double meanDistance, long double eccentricity, long double inclination, int colour) : StellarObject(name, 4, radius, mass, meanDistance, eccentricity, inclination, colour){
     
 }
