+++ universim2/positionVector.hpp
@@ -7,40 +7,40 @@ class Matrix3d;
 
 class PositionVector{
 private:
-    double x;
-    double y;
-    double z;
-    double length;
+    long double x;
+    long double y;
+    long double z;
+    long double length;
     bool validLength = false;
     std::string vectorString;
 
 public:
     PositionVector();
-    PositionVector(double x, double y, double z);
+    PositionVector(long double x, long double y, long double z);
     PositionVector(PositionVector *parent);
-    double getX();
-    double getY();
-    double getZ();
-    double getLength();
-    void setX(double x);
-    void setY(double y);
-    void setZ(double z);
-    double distance(PositionVector position);
+    long double getX();
+    long double getY();
+    long double getZ();
+    long double getLength();
+    void setX(long double x);
+    void setY(long double y);
+    void setZ(long double z);
+    long double distance(PositionVector position);
     // Returns the normalised value of the vector, without changing the vector itself
     PositionVector normalise();
     PositionVector operator + (PositionVector position);
     PositionVector operator - (PositionVector position);
-    PositionVector operator * (double multiplier);
-    PositionVector operator / (double multiplier);
+    PositionVector operator * (long double multiplier);
+    PositionVector operator / (long double multiplier);
     void operator = (PositionVector position);
     void operator += (PositionVector position);
     void operator -= (PositionVector position);
-    void operator *= (double multiplier);
-    void operator /= (double multiplier);
+    void operator *= (long double multiplier);
+    void operator /= (long double multiplier);
     bool operator == (PositionVector position);
 
     // Functions for the computation of multipole expansion
-    double rQr(Matrix3d Q);
+    long double rQr(Matrix3d Q);
     PositionVector Qr(Matrix3d Q);
 
     const char *toString();
